feat(agents): native multi-agent governance M1a-M4 + closure + dashboard
- agents/: identity (DID chain + admin trust root + Sybil), task (dual-state
  lifecycle, idempotent re-claim), bus (at-most-one lease queue), roles
  (5-role whitelists), planner (schema-first decomposition), runtime (thin
  loop; real toolchains for all 5 roles; optional signer so B2 trust checks
  verify runtime-produced events), trust (B4 org diversity + weak-signal
  reputation)
- trust_model: B4 diversity activated via provider, reputation floor
- surfaces: REST/CLI/tools/MCP for agents+tasks+runtime+trust; meta
  endpoints export task transitions + role whitelists (single source of
  truth); admin JWT issuance (ADMIN_USERNAME/ADMIN_PASSWORD)
- dashboard: auth + admin validate panels, Agents/Tasks/Runtime/Trust
  pages, meta-driven task actions
- fix: chain_kind was fed chain keys (task.py); _extract_org handled
  normalized DIDDocument (no .get); _mem_store store_document -> store
- tests: 1731 fast-path green, zero regression (was 1638)

+++ experiments/e33_challenge_game.py
@@ -202,12 +202,12 @@ def _interpolate_f_star(rows: list[dict]) -> float | None:
     for row in rows:
         if row["lazy_payoff_per_round"] <= E_HONEST:
             if prev is None:
-                return row["f"]
+                return float(row["f"])
             f0, y0 = prev["f"], prev["lazy_payoff_per_round"]
             f1, y1 = row["f"], row["lazy_payoff_per_round"]
             if y1 == y0:
-                return f1
-            return round(f0 + (E_HONEST - y0) * (f1 - f0) / (y1 - y0), 4)
+                return float(f1)
+            return float(round(f0 + (E_HONEST - y0) * (f1 - f0) / (y1 - y0), 4))
         prev = row
     return None
